+++ Source/Core Providers/Core_vJoyInterfaceWrap/Core_vJoyInterfaceWrap.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using HidWizards.IOWrapper.Core.Devcon;
 using HidWizards.IOWrapper.DataTransferObjects;
 using HidWizards.IOWrapper.ProviderInterface.Interfaces;
 
@@ -14,32 +15,37 @@ namespace Core_vJoyInterfaceWrap
     [Export(typeof(IProvider))]
     public class Core_vJoyInterfaceWrap : IOutputProvider
     {
-        public bool IsLive { get { return isLive; } }
-        private bool isLive = true; // ToDo: Detect if driver is present and report as appropriate
+        public bool IsLive => _isLive;
+        private readonly bool _isLive = false;
 
         bool disposed;
         public static vJoyInterfaceWrap.vJoy vJ = new vJoyInterfaceWrap.vJoy();
-        private VjoyDevice[] vJoyDevices = new VjoyDevice[16];
-        private Dictionary<Guid, uint> subscriptionToDevice = new Dictionary<Guid, uint>();
-        static private Dictionary<int, string> axisNames = new Dictionary<int, string> { { 0, "X" }, { 1,"Y" }, { 2, "Z" }, { 3, "Rx" }, { 4, "Ry" }, { 5, "Rz" }, { 6, "Sl0" }, { 7, "Sl1" } };
-        static private List<BindingReport>[] povBindingInfos = new List<BindingReport>[4];
+        private VjoyDevice[] _vJoyDevices = new VjoyDevice[16];
+        private readonly Dictionary<Guid, uint> _subscriptionToDevice = new Dictionary<Guid, uint>();
+        private static readonly Dictionary<int, string> AxisNames = new Dictionary<int, string> { { 0, "X" }, { 1,"Y" }, { 2, "Z" }, { 3, "Rx" }, { 4, "Ry" }, { 5, "Rz" }, { 6, "Sl0" }, { 7, "Sl1" } };
+        private static readonly List<BindingReport>[] PovBindingInfos = new List<BindingReport>[4];
+        private static Guid InterfaceGuid => Guid.Parse("{781EF630-72B2-11d2-B852-00C04FAD5101}");
+        private readonly string _errorMessage;
 
-        private List<DeviceReport> deviceReports;
+        private List<DeviceReport> _deviceReports;
 
         public Core_vJoyInterfaceWrap()
         {
+            _isLive = Devcon.FindDeviceByInterfaceId(InterfaceGuid, out var path, out _);
+            if (_isLive) _errorMessage = string.Empty;
+            else _errorMessage = "vJoy device driver not installed";
             for (uint i = 0; i < 16; i++)
             {
-                vJoyDevices[i] = new VjoyDevice(i + 1);
+                _vJoyDevices[i] = new VjoyDevice(i + 1);
             }
-            for (int p = 0; p < 4; p++)
+            for (var p = 0; p < 4; p++)
             {
-                povBindingInfos[p] = new List<BindingReport>();
-                for (int d = 0; d < 4; d++)
+                PovBindingInfos[p] = new List<BindingReport>();
+                for (var d = 0; d < 4; d++)
                 {
-                    povBindingInfos[p].Add(new BindingReport
+                    PovBindingInfos[p].Add(new BindingReport
                     {
-                        Title = povDirections[d],
+                        Title = PovDirections[d],
                         Category = BindingCategory.Momentary,
                         BindingDescriptor = new BindingDescriptor
                         {
@@ -70,9 +76,9 @@ namespace Core_vJoyInterfaceWrap
             {
                 for (uint devId = 0; devId < 16; devId++)
                 {
-                    vJoyDevices[devId].Dispose();
+                    _vJoyDevices[devId].Dispose();
                 }
-                vJoyDevices = null;
+                _vJoyDevices = null;
                 vJ = null;
             }
             disposed = true;
@@ -100,14 +106,15 @@ namespace Core_vJoyInterfaceWrap
                 {
                     ProviderName = ProviderName
                 },
-                Devices = deviceReports
+                Devices = _deviceReports,
+                ErrorMessage = _errorMessage
             };
             return pr;
         }
 
         public DeviceReport GetOutputDeviceReport(DeviceDescriptor deviceDescriptor)
         {
-            foreach (var deviceReport in deviceReports)
+            foreach (var deviceReport in _deviceReports)
             {
                 if (deviceReport.DeviceDescriptor.DeviceHandle == deviceDescriptor.DeviceHandle && deviceReport.DeviceDescriptor.DeviceInstance == deviceDescriptor.DeviceInstance)
                 {
@@ -125,23 +132,23 @@ namespace Core_vJoyInterfaceWrap
         public bool SubscribeOutputDevice(OutputSubscriptionRequest subReq)
         {
             var devId = DevIdFromHandle(subReq.DeviceDescriptor.DeviceHandle);
-            vJoyDevices[devId].Add(subReq);
-            subscriptionToDevice.Add(subReq.SubscriptionDescriptor.SubscriberGuid, devId);
+            _vJoyDevices[devId].Add(subReq);
+            _subscriptionToDevice.Add(subReq.SubscriptionDescriptor.SubscriberGuid, devId);
             return true;
         }
 
         public bool UnSubscribeOutputDevice(OutputSubscriptionRequest subReq)
         {
-            uint devId = subscriptionToDevice[subReq.SubscriptionDescriptor.SubscriberGuid];
-            vJoyDevices[devId].Remove(subReq);
-            subscriptionToDevice.Remove(subReq.SubscriptionDescriptor.SubscriberGuid);
+            uint devId = _subscriptionToDevice[subReq.SubscriptionDescriptor.SubscriberGuid];
+            _vJoyDevices[devId].Remove(subReq);
+            _subscriptionToDevice.Remove(subReq.SubscriptionDescriptor.SubscriberGuid);
             return true;
         }
 
         public bool SetOutputState(OutputSubscriptionRequest subReq, BindingDescriptor bindingDescriptor, int state)
         {
-            var devId = subscriptionToDevice[subReq.SubscriptionDescriptor.SubscriberGuid];
-            if (!vJoyDevices[devId].IsAcquired)
+            var devId = _subscriptionToDevice[subReq.SubscriptionDescriptor.SubscriberGuid];
+            if (!_vJoyDevices[devId].IsAcquired)
             {
                 return false;
             }
@@ -157,7 +164,7 @@ namespace Core_vJoyInterfaceWrap
                     int pov = (int)(Math.Floor((decimal)(bindingDescriptor.Index / 4)));
                     int dir = bindingDescriptor.Index % 4;
                     //Log("vJoy POV output requested - POV {0}, Dir {1}, State {2}", pov, dir, state);
-                    vJoyDevices[devId].SetPovState(pov, dir, state);
+                    _vJoyDevices[devId].SetPovState(pov, dir, state);
                     break;
 
                 default:
@@ -179,7 +186,7 @@ namespace Core_vJoyInterfaceWrap
 
         private void QueryDevices()
         {
-            deviceReports = new List<DeviceReport>();
+            _deviceReports = new List<DeviceReport>();
             for (uint i = 0; i < 16; i++)
             {
                 var id = i + 1;
@@ -188,7 +195,7 @@ namespace Core_vJoyInterfaceWrap
                     var handle = i.ToString();
                     var device = new DeviceReport
                     {
-                        DeviceName = String.Format("vJoy Stick {0}", id),
+                        DeviceName = $"vJoy Stick {id}",
                         DeviceDescriptor = new DeviceDescriptor
                         {
                             DeviceHandle = handle
@@ -200,14 +207,14 @@ namespace Core_vJoyInterfaceWrap
                         Title = "Axes"
                     };
 
-                    int axisCount = 0;
-                    for (int ax = 0; ax < 8; ax++)
+                    var axisCount = 0;
+                    for (var ax = 0; ax < 8; ax++)
                     {
                         if (vJ.GetVJDAxisExist(id, AxisIdToUsage[ax]))
                         {
                             axisNode.Bindings.Add(new BindingReport
                             {
-                                Title = axisNames[ax],
+                                Title = AxisNames[ax],
                                 Category = BindingCategory.Signed,
                                 BindingDescriptor = new BindingDescriptor
                                 {
@@ -256,30 +263,30 @@ namespace Core_vJoyInterfaceWrap
                             Title = "POVs"
                         };
 
-                        for (int p = 0; p < povCount; p++)
+                        for (var p = 0; p < povCount; p++)
                         {
                             var povNode = new DeviceReportNode
                             {
                                 Title = "POV #" + (p + 1)
                             };
-                            povNode.Bindings = povBindingInfos[p];
+                            povNode.Bindings = PovBindingInfos[p];
                             povsNode.Nodes.Add(povNode);
                         }
                         device.Nodes.Add(povsNode);
                     }
 
-                    deviceReports.Add(device);
+                    _deviceReports.Add(device);
                 }
             }
 
         }
 
-        private uint DevIdFromHandle(string handle)
+        private static uint DevIdFromHandle(string handle)
         {
             return Convert.ToUInt32(handle);
         }
 
-        private static List<HID_USAGES> AxisIdToUsage = new List<HID_USAGES>
+        private static readonly List<HID_USAGES> AxisIdToUsage = new List<HID_USAGES>
         {
             HID_USAGES.HID_USAGE_X, HID_USAGES.HID_USAGE_Y, HID_USAGES.HID_USAGE_Z,
             HID_USAGES.HID_USAGE_RX, HID_USAGES.HID_USAGE_RY, HID_USAGES.HID_USAGE_RZ,
@@ -287,35 +294,35 @@ namespace Core_vJoyInterfaceWrap
 
         private class VjoyDevice
         {
-            public bool IsAcquired { get { return acquired; } }
+            public bool IsAcquired { get { return _acquired; } }
 
-            private uint deviceId;
-            private bool acquired;
-            private Dictionary<Guid, OutputSubscriptionRequest> subReqs = new Dictionary<Guid, OutputSubscriptionRequest>();
-            private POVHandler[] povHandlers = new POVHandler[4];
+            private readonly uint _deviceId;
+            private bool _acquired;
+            private readonly Dictionary<Guid, OutputSubscriptionRequest> _subReqs = new Dictionary<Guid, OutputSubscriptionRequest>();
+            private readonly POVHandler[] _povHandlers = new POVHandler[4];
 
             public VjoyDevice(uint id)
             {
-                deviceId = id;
+                _deviceId = id;
                 for (uint i = 0; i < 4; i++)
                 {
-                    povHandlers[i] = new POVHandler(deviceId, i + 1);
+                    _povHandlers[i] = new POVHandler(_deviceId, i + 1);
                 }
             }
 
             public void Add(OutputSubscriptionRequest subReq)
             {
-                if (subReqs.Count == 0)
+                if (_subReqs.Count == 0)
                 {
                     Acquire();
                 }
-                subReqs.Add(subReq.SubscriptionDescriptor.SubscriberGuid, subReq);
+                _subReqs.Add(subReq.SubscriptionDescriptor.SubscriberGuid, subReq);
             }
 
             public void Remove(OutputSubscriptionRequest subReq)
             {
-                subReqs.Remove(subReq.SubscriptionDescriptor.SubscriberGuid);
-                if (subReqs.Count == 0)
+                _subReqs.Remove(subReq.SubscriptionDescriptor.SubscriberGuid);
+                if (_subReqs.Count == 0)
                 {
                     Relinquish();
                 }
@@ -323,7 +330,7 @@ namespace Core_vJoyInterfaceWrap
 
             public void SetPovState(int pov, int dir, int state)
             {
-                povHandlers[pov].SetState(dir, state);
+                _povHandlers[pov].SetState(dir, state);
             }
 
             public void Dispose()
@@ -333,34 +340,34 @@ namespace Core_vJoyInterfaceWrap
 
             private void Acquire()
             {
-                if (!acquired)
+                if (!_acquired)
                 {
-                    vJ.AcquireVJD(deviceId);
-                    acquired = true;
-                    Log("Acquired vJoy device {0}", deviceId);
+                    vJ.AcquireVJD(_deviceId);
+                    _acquired = true;
+                    Log("Acquired vJoy device {0}", _deviceId);
                 }
             }
 
             private void Relinquish()
             {
-                if (acquired)
+                if (_acquired)
                 {
-                    vJ.RelinquishVJD(deviceId);
-                    acquired = false;
-                    Log("Relinquished vJoy device {0}", deviceId);
+                    vJ.RelinquishVJD(_deviceId);
+                    _acquired = false;
+                    Log("Relinquished vJoy device {0}", _deviceId);
                 }
             }
 
             private class POVHandler
             {
-                private int[] axes = { 0, 0 };
-                private uint deviceId;
-                private uint povId;
+                private readonly int[] _axes = { 0, 0 };
+                private readonly uint _deviceId;
+                private readonly uint _povId;
 
                 public POVHandler(uint device, uint pov)
                 {
-                    deviceId = device;
-                    povId = pov;
+                    _deviceId = device;
+                    _povId = pov;
                 }
 
                 public void SetState(int dir, int state)
@@ -368,7 +375,7 @@ namespace Core_vJoyInterfaceWrap
                     var axis = DirToAxis(dir);
                     var vector = DirToCoordinate(dir);
                     //Log("POV Dir: {0} Axis: {1}, Vector: {2}", dir, axis, vector);
-                    axes[axis] = state == 0 ? 0 : vector;
+                    _axes[axis] = state == 0 ? 0 : vector;
                     SetPovState();
                 }
 
@@ -381,9 +388,9 @@ namespace Core_vJoyInterfaceWrap
 
                 private void SetPovState()
                 {
-                    var angle = (axes[0] == 0 && axes[1] == 0 ? -1 : GetAngle());
+                    var angle = (_axes[0] == 0 && _axes[1] == 0 ? -1 : GetAngle());
                     //Log("Pov Angle: {0}", angle);
-                    vJ.SetContPov(angle, deviceId, povId);
+                    vJ.SetContPov(angle, _deviceId, _povId);
                 }
 
                 private int DirToAxis(int dir)
@@ -393,7 +400,7 @@ namespace Core_vJoyInterfaceWrap
 
                 private int GetAngle()
                 {
-                    int angle = (int)(Math.Atan2(axes[0], axes[1]) * (180 / Math.PI)) * 100;
+                    int angle = (int)(Math.Atan2(_axes[0], _axes[1]) * (180 / Math.PI)) * 100;
                     if (angle < 0)
                         angle = 36000 + angle;
                     return angle;
@@ -402,6 +409,6 @@ namespace Core_vJoyInterfaceWrap
 
         }
 
-        private static List<string> povDirections = new List<string> { "Up", "Right", "Down", "Left" };
+        private static readonly List<string> PovDirections = new List<string> { "Up", "Right", "Down", "Left" };
     }
 }