+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [MIDI provider] IsLive is always true, as MIDI is always present
 - [TitanOne Provider] IsLive now reflects connected status of device
 - [TitanOne Provider] Reports 0 devices if IsLive is false
+- [vJoy Provider] IsLive now reflects whether driver is installed
 ### Changed 
 ### Deprecated
 ### Removed